extract constants and add  annotations
