+++ src/main/java/edu/kis/vh/nursery/DefaultCountingOutRhymer.java
@@ -1,0 +1,36 @@
+package edu.kis.vh.nursery;
+
+public class DefaultCountingOutRhymer {
+
+    public static final int SIZE = 12;
+    public static final int EMPTY = -1;
+    public static final int FULL = 11;
+    public static final int ERROR_CODE = -1;
+    private final int[] numbers = new int[SIZE];
+
+    public int total = EMPTY;
+
+    public void countIn(int in) {
+        if (!isFull())
+            numbers[++total] = in;
+    }
+
+    public boolean callCheck() {
+        return total == EMPTY;
+    }
+
+    public boolean isFull() {
+        return total == FULL;
+    }
+
+    protected int peekaboo() {
+        if (callCheck()) return ERROR_CODE;
+        return numbers[total];
+    }
+
+    public int countOut() {
+        if (callCheck())
+            return ERROR_CODE;
+        return numbers[total--];
+    }
+}
